+++ src/constants.js
@@ -27,6 +27,8 @@ const DOM_ELEMENT_POSITION_KEYS = [
 const DOM_ELEMENT_SIZE_KEYS = [
   'clientHeight',
   'clientWidth',
+  'naturalHeight',
+  'naturalWidth',
   'offsetHeight',
   'offsetWidth',
   'scrollHeight',