+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # remeasure changelog
 
+#### 1.2.0
+* Component now only updates when one of the requested keys updates (or if used in standard way, when any key updates)
+* Add `naturalHeight` and `naturalWidth` properties
+  * `img` is the only element that supports this natively, for all other elements these values reflect `scrollHeight` and `scrollWidth`
+
+#### 1.1.0
+* Add `clientLeft`, `clientTop`, `scrollLeft`, and `scrollTop` properties
+
+#### 1.0.1
+* Small changes to README
+
 #### 1.0.0
 * Initial release